Branch , Bonus , Branch Manager services changed to orm

+++ Backend/Models/BranchManagerUpdaterModel.cs
@@ -5,7 +5,7 @@ namespace Backend.Models
     public class BranchManagerUpdaterModel : UserUpdaterModel
     {
         public int? Branch_Manager_ID { get; set; }
-        public int? Salary { get; set; }
+        public decimal? Salary { get; set; }
         public int? Penalties { get; set; }
         public int? Bonuses { get; set; }
         public DateOnly? Hire_Date { get; set; }
